chore: HomeView 목데이터 추가
목데이터 추가하였습니다.
